+++ examples/car/run.py
@@ -1,15 +1,20 @@
 from pprint import pprint
 from .shared import get_model
 
-# use_results = "snapshot"
+# caching_level = "incrementally"
+caching_level = "never"
+# use_results = "cached"
 use_results = "none"
 
 hmodel1 = get_model("model1.yml")
+hmodel1.set_caching_level(caching_level)
+hmodel1.clear_cache()
 query = ["cars[0].price", "cars[1].price"]
 response = hmodel1.get(query, use_results=use_results)
 print(response)
 
 hmodel2 = get_model("model2.yml")
+hmodel2.set_caching_level(caching_level)
 query = [
     "cars[:].parts[:].price",  # price for all components for all cars
     "cars[:].price",  # price for all cars
@@ -19,6 +24,7 @@ pprint(response)
 
 
 hmodel3 = get_model("model3.yml")
+hmodel3.set_caching_level(caching_level)
 query = ["cars[:].price"]  # price for all cars
 response = hmodel3.get(query, use_results=use_results)
 print(response)